[expo-updates][android] Remove legacy manifest support
