+++ packages/expo-manifests/android/src/main/java/expo/modules/manifests/core/Manifest.kt
@@ -244,9 +244,6 @@ abstract class Manifest(protected val json: JSONObject) {
   companion object {
     @JvmStatic fun fromManifestJson(manifestJson: JSONObject): Manifest {
       return when {
-        manifestJson.has("releaseId") -> {
-          LegacyManifest(manifestJson)
-        }
         manifestJson.has("metadata") -> {
           NewManifest(manifestJson)
         }
